добавление хука useFavorite
